fix AdminDeleteMenuItemButton

+++ src/app/components/admin/AdminDeleteMenuItemButton.tsx
@@ -1,50 +1,67 @@
 'use client'
 
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/navigation'
+import { useSession } from 'next-auth/react'
 import { twMerge } from 'tailwind-merge'
+import { handleToast } from '@/utils/handleToast'
 
-const AdminDeleteMenuItemButton = () => {
+type Props = {
+  menuItemCategory: string
+  menuItemId: string
+}
+
+const AdminDeleteMenuItemButton = ({ menuItemCategory, menuItemId }: Props) => {
   const router = useRouter()
-  const queryClient = useQueryClient()
   const { data: session } = useSession()
   const isAdmin = session?.user?.isAdmin
 
   const handleDelete = async () => {
-    console.log('handleDelete')
-    const menuItemId = 'router'
+    if (!isAdmin) return
 
-    deleteMenuItem.mutate(menuItemId)
-  }
-
-  const deleteMenuItem = useMutation({
-    mutationFn: async (menuItemId: string) => {
+    try {
       const response = await fetch(
-        `${process.env.NEXTAUTH_URL}/menuItems/${menuItemId}`,
+        `${process.env.NEXT_PUBLIC_NEXTAUTH_URL}/api/menuItems/${menuItemId}`,
         {
           method: 'DELETE',
           headers: {
             'Content-Type': 'application/json',
-            cache: 'no-store',
           },
         }
       )
-      return response.json()
-    },
-    // onSuccess: () => {
-    //   queryClient.invalidateQueries('MenuItems')
-    // },
-  })
+
+      if (!response.ok) {
+        throw new Error()
+      }
+      const data = await response.json()
+
+      // notify user
+      handleToast({
+        type: 'success',
+        message: `${data.message}.`,
+      })
+
+      // navigate to menu/[category] page
+      router.push(`/menu/${menuItemCategory}`)
+    } catch (error) {
+      console.log(error)
+      handleToast({
+        type: 'error',
+        message: `Error deleting menu item.`,
+      })
+    }
+  }
 
   if (!isAdmin) return null
 
   return (
-    <div className='absolute right-1 top-[6.5rem]'>
+    <div className='absolute bottom-1 right-1'>
       <button
         onClick={handleDelete}
         type='submit'
-        className={twMerge('btn', 'px-4 py-2 text-base')}
+        className={twMerge(
+          'btn',
+          'z-50 bg-tertiary-dark px-4 py-2 text-base transition duration-300 ease-in-out hover:scale-105'
+        )}
       >
         Delete
       </button>
